pyproject.toml
[project]
name = "kraxcli"
dynamic=["version"]
description = "Wrapper for pyserial-ports/mpremote/upydev"
authors = [
  {name="Василий", email="vlinnik@mail.ru"}
]
dependencies = ["colorlog","telnetlib3","upydev","upydevice","dill","pyserial","mpremote","argparse","questionary"]

[build-system]
requires = ["setuptools>=42", "wheel","setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
version_scheme = "post-release"
write_to = "src/kraxcli/_version.py"

[project.scripts]
kraxcli = "kraxcli.__main__:main"
